+++ phylo/src/optimisers/blen_optimiser_tests.rs
@@ -3,12 +3,12 @@ use std::path::PathBuf;
 use approx::assert_relative_eq;
 
 use crate::evolutionary_models::DNAModelType;
-use crate::optimisers::branch_length_optimiser::BranchOptimiser;
-use crate::optimisers::PhyloOptimiser;
+use crate::optimisers::{BranchOptimiser, PhyloOptimiser};
 use crate::phylo_info::PhyloInfoBuilder;
 use crate::pip_model::{PIPCost, PIPDNAModel};
-use crate::substitution_models::dna_models::DNASubstModel;
-use crate::substitution_models::{SubstLikelihoodCost, SubstitutionModel};
+use crate::substitution_models::{
+    dna_models::DNASubstModel, SubstLikelihoodCost, SubstitutionModel,
+};
 use crate::tree::tree_parser::from_newick_string;
 use crate::tree::NodeIdx::Internal as Int;
